Removed internal retry protection against invalid markup.
This code was added beased on a flawed premise that thought some pages
were emitting invalid markup sporadically. In fact, invalid markup was
consistent for Valve Index hardware page (#1059530) and the temporary
logging facility revealed it would always retry 5 times.

This appeared to "fix" the issue because retrying promoted the error to
FailingTooHardException which was handled, whereas the underlying
InvalidMarkupException was not handled by the Steam Importer. The
correct fix should be to just handle the InvalidMarkupException rather
than implement retries in this library.