Fixed game reviews parsing.
On 2024/2/7 Valve changed the icon associated with Steam reviews
depending on whether the review is "counter" or not. Previously, this
icon was displayed as either the Steam logo (for Steam purchases) or a
key icon (for Steam keys). These have been changed to stars, either
filled (Steam/free license) or hollow (key/gift/temporary license).
Although the definition appears to have changed, the functionality
appears identical to previously, with only the icons being different.